version 2.4 : debian is now with sudo
dropped sudo-less version of sudo for good.
debian and all derivatives must now have sudo installed and configured to run this script, if they are in the 0.0000001% that don't / don't want to, I really cannot be asked anymore. this has been pointlessly nightmarish to support and debian just dropped support for new windows of gnome-terminal being opened as root, which killed the old version of my script.